Separate tests on patch size and patch spreading
